day7 part2: Make WAY more efficient
if C is the number of crabs, and P is the number of positions, the time
complexities of the old and new algorithms are:

Old algorithm: O(C*P). For each position, iterate through the crabs and
add up the fuel required for each one.

New algorithm: O(P). Iterate through the positions backwards then
forwards, to get accumulated fuel required for all crabs from each
direction (for each position). Then take the min of backwards + forwards
out of all the positions.